First round of mask refactor.
